Add volatility comparison project and main menu
